Adds workaround for bug in esp32 HCI_SUBEVENT_LE_CONNECTION_COMPLETE event.
After calling connect_cancel a HCI_SUBEVENT_LE_CONNECTION_COMPLETE event shall
be generated if currently connecting. Problem is that this event after a
cancel_connect doesn't contain a mac address, just 0's. This causes the btstack
to not remove this connection, and the next time connecting to this address, the
btstack will beleive that it's already connected to this address.