a2dp source: fix crash on outgoing connection after previous incoming one
The pointer to the user media codec configuration buffer was stored in stream_endpoint.remote_configuration, which was set to zero in avdtp_reset_stream_endpoint. The user buffer is now stored separately and it is updateds when the configuration gets set/updated. The field media_codec_sbc_info is used by a2dp configure and reconfigure.