esp32: update CMakeLists.txt file
This commit updates the esp32 CMakeLists.txt so that it is compatible
with the new way of creating CMakeLists.txt files in ESP-IDF.

Basically it replaces the old register_component() with
idf_component_register().

It also removes the relative references to lwip. This is because some
projects might want to puse btstack directly in their private "components"
folder (instead of being placed in the esp-idf/component folder)

It has been tested both by:
 - placing btstack in esp-idf/components
 - placing btstack in my_own_project/components

 and it compiled Ok using "idf.py build"

 Tested "pan_lwip_http_server" and compiled Ok as well.