3rd-party/lwip/dhserver: interface for dhcp reply
If more than one interface is used, chances are that the last one added
will have changed the default interface used for sending data (see
netif_set_default). This causes problems for the DHCP server in the PAN
example as the DHCP reply is sent on the wrong network. On Windows this
can cause it to use the wrong address.

Send replies to the interface that received the dhcp request.