Buffer not always big enough for HID GET_REPORT handling.
NOTE: This may use a lot of stack space, so perhaps it is better to allocate the report buffer on the heap instead.

The buffer used to get the GET_REPORT data into from the hci_device_get_report() callback is fixed at 0x30 bytes. This could be too small to contain the full data copied in by the hci_device_get_report() buffer. The actual size copied in comes from the HID report descriptor which is custom for each HID device. The hci_device_get_report() callback could potentially perform a buffer overflow on the stack when copying in its data. This of course has security implications for potential remote code execution.

Not only can it overflow this buffer, it will also not return the full contents to the host as size checks are done outside of the callback and will only ever return a max of the current buffer size which is 0x30 bytes.

From the Bluetooth HID Profile doc:
5.2.4 Bluetooth HID device L2CAP Requirements
5.2.4.1 Bluetooth HID device MTU Usage
The implementation of L2CAP on the Bluetooth HID device shall support a receive MTU which is large enough to accommodate the largest report (input, output or feature) defined in the Bluetooth HID device’s report descriptor. This means the MTU on both the interrupt and Control channels shall be large enough to accommodate the largest packet possible, either a SET_REPORT on the Control channel or DATA on the Interrupt channel, including all overheads such as the transaction header and report ID.