hci: fixing two easy fuzz bugs: 0 sco packet length and connection use-after-free
Bug 1: sco packet length
The SCO packet length can be zero when connections are not yet initialized.
Make sure that we don't trigger a division-by-zero bug during this initial
stage by setting the SCO packet length to a reasonable minimum.

Bug 2: connection user-after-free
The connection is not reset correctly during fuzzing, resulting in a
potential use-after-free in the fuzzer code. Quick fix to correctly
reset the timers by stopping them.